08.19 - 4

+++ fin_the_pen/src/main/java/project/fin_the_pen/data/member/UserResponseDTO.java
@@ -6,7 +6,7 @@ import java.util.Date;
 
 @Data
 public class UserResponseDTO {
-    private int id;
+    private Long id;
     private String user_id;
     private String name;
     private Date baby;
@@ -14,7 +14,7 @@ public class UserResponseDTO {
     private String userRole;
     private String phone_number;
 
-    public UserResponseDTO(int id, String user_id, String name, Date baby, Date registerDate, String userRole, String phone_number) {
+    public UserResponseDTO(Long id, String user_id, String name, Date baby, Date registerDate, String userRole, String phone_number) {
         this.id = id;
         this.user_id = user_id;
         this.name = name;
